added support for skip test cases
