+++ tests/pyproject.toml
@@ -24,7 +24,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
-addopts = "-p no:warnings -s -v --alluredir=tests/allure/results"
+addopts = "-p no:warnings -s -v --alluredir=tests/allure/results --clean-alluredir"
 markers = [
     # tests used in cloudbeat CI
     "pre_merge",
